feat(proxy): route web requests by configured header
Add optional header-based segment selection for regular, chunked, and websocket lifecycle messages while preserving existing routing when the header is absent.

Handle Integer.MIN_VALUE safely in ConsistentHashing without remapping other hash values.

+++ common/src/main/java/io/fluxzero/common/ConsistentHashing.java
@@ -75,7 +75,7 @@ public class ConsistentHashing {
      * @return the segment index
      */
     public static int computeSegment(String routingKey, Function<String, Integer> hashFunction, int maxSegments) {
-        return Math.abs(hashFunction.apply(routingKey)) % maxSegments;
+        return (int) (Math.abs((long) hashFunction.apply(routingKey)) % maxSegments);
     }
 
     /**
